requirements.txt
certifi==2016.2.28
Django==2.2
django-environ==0.4.5
djangorestframework==3.9.2
Markdown==3.1
pytz==2019.1
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.24.1
wincertstore==0.2
